Hosted by section added and powered from Firebase, also event detail screen refactored and separated by widgets
