+++ lib/src/models/event.dart
@@ -13,6 +13,7 @@ class Event {
   final String calendarUrl;
   final String locationUrl;
   final List<String> attendees;
+  final List<dynamic> speakers;
 
   int get attendeesCount => attendees.length;
 
@@ -29,6 +30,7 @@ class Event {
     required this.calendarUrl,
     required this.locationUrl,
     required this.attendees,
+    required this.speakers,
   });
 
   factory Event.fromJson({
@@ -37,8 +39,7 @@ class Event {
   }) {
     final dateTimeTimeStamp = json['dateTime'] as Timestamp;
 
-    final recommendations =
-        (json['recommendations'] as String ?? '').replaceAll("\\n", "\n");
+    final recommendations = (json['recommendations']).replaceAll("\\n", "\n");
 
     final attendees = (json['attendees'] as List<dynamic>? ?? [])
         .map<String>((uid) => '$uid')
@@ -57,6 +58,7 @@ class Event {
       calendarUrl: json['calendar_url'] ?? '',
       recommendations: recommendations,
       attendees: attendees,
+      speakers: json['speakers'] ?? {},
     );
   }
 }
